Reorganize code base structure
