+++ config.go
@@ -3,10 +3,10 @@ package main
 import (
 	"time"
 
-	"github.com/iotaledger/evil-tools/accountwallet"
-	"github.com/iotaledger/evil-tools/evilwallet"
+	"github.com/iotaledger/evil-tools/pkg/accountwallet"
+	"github.com/iotaledger/evil-tools/pkg/evilwallet"
+	"github.com/iotaledger/evil-tools/pkg/spammer"
 	"github.com/iotaledger/evil-tools/programs"
-	"github.com/iotaledger/evil-tools/spammer"
 )
 
 // Nodes used during the test, use at least two nodes to be able to double spend.
