+++ parse.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/iotaledger/evil-tools/accountwallet"
-	"github.com/iotaledger/evil-tools/evilwallet"
+	"github.com/iotaledger/evil-tools/pkg/accountwallet"
+	"github.com/iotaledger/evil-tools/pkg/evilwallet"
 	"github.com/iotaledger/hive.go/ierrors"
 )
