+++ pkg/accountwallet/faucet.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/mr-tron/base58"
 
-	"github.com/iotaledger/evil-tools/models"
-	"github.com/iotaledger/evil-tools/utils"
+	"github.com/iotaledger/evil-tools/pkg/models"
+	"github.com/iotaledger/evil-tools/pkg/utils"
 	"github.com/iotaledger/hive.go/ierrors"
 	"github.com/iotaledger/hive.go/lo"
 	"github.com/iotaledger/iota-core/pkg/blockhandler"
