+++ pkg/evilwallet/customscenarios.go
