+++ pkg/evilwallet/evilscenario.go
