+++ pkg/interactive/menu.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/iotaledger/evil-tools/evilwallet"
+	"github.com/iotaledger/evil-tools/pkg/evilwallet"
 )
 
 // region Printer /////////////////////////////////////////////////////////////////////////////////////////////////////////
