+++ pkg/models/connector.go
