+++ pkg/spammer/errors.go
